+++ commands/compile.ts
@@ -1,5 +1,7 @@
 import vscode = require('vscode');
 import fs = require('fs');
+import path = require('path');
+import glob = require('glob');
 import { AtelierAPI } from '../api';
 import { currentFile, CurrentFile, outputChannel } from '../utils';
 import { documentContentProvider, config } from '../extension';
@@ -13,7 +15,7 @@ async function compileFlags(): Promise<string> {
   });
 }
 
-async function importFile(file: CurrentFile, flags: string): Promise<any> {
+async function importFile(file: CurrentFile): Promise<any> {
   const api = new AtelierAPI();
   return api
     .putDoc(
@@ -23,13 +25,7 @@ async function importFile(file: CurrentFile, flags: string): Promise<any> {
         content: file.content.split(/\r?\n/)
       },
       true
-    )
-    .then(() => compile(file, flags))
-    .catch((error: Error) => {
-      outputChannel.appendLine(error.message);
-      outputChannel.show(true);
-      vscode.window.showErrorMessage(error.message);
-    });
+    );
 }
 
 function updateOthers(others: string[]) {
@@ -39,30 +35,42 @@ function updateOthers(others: string[]) {
   });
 }
 
-async function loadChanges(file: CurrentFile): Promise<any> {
+async function loadChanges(files: CurrentFile[]): Promise<any> {
   const api = new AtelierAPI();
-  return api.getDoc(file.name).then(data => {
-    fs.writeFileSync(file.fileName, (data.result.content || []).join('\n'));
-    api
-      .actionIndex([file.name])
-      .then(data => data.result.content[0].others)
-      .then(updateOthers);
-  });
+  return Promise.all(files.map(file =>
+    api.getDoc(file.name).then(data => {
+      fs.writeFileSync(file.fileName, (data.result.content || []).join('\n'));
+      return api
+        .actionIndex([file.name])
+        .then(data => data.result.content[0].others)
+        .then(updateOthers);
+    })
+  ));
 }
 
-async function compile(file: CurrentFile, flags: string): Promise<any> {
+async function compile(docs: CurrentFile[], flags?: string): Promise<any> {
+  flags = flags || config().compileFlags;
   const api = new AtelierAPI();
   return api
-    .actionCompile([file.name], flags)
+    .actionCompile(docs.map(el => el.name), flags)
     .then(data => {
+      let info = docs.length > 1 ? '' : `${docs[0].name}: `;
       if (data.status && data.status.errors && data.status.errors.length) {
-        throw new Error(`${file.name}: Compile error`);
+        throw new Error(`${info}Compile error`);
       } else {
-        vscode.window.showInformationMessage(`${file.name}: Compile successed`);
+        vscode.window.showInformationMessage(`${info}Compile successed`, 'Hide');
       }
-      return file;
+      return docs;
     })
-    .then(loadChanges);
+    .then(loadChanges)
+    .catch((error: Error) => {
+      outputChannel.appendLine(error.message);
+      outputChannel.show(true);
+      vscode.window.showErrorMessage(error.message, 'Show details')
+        .then(data => {
+          outputChannel.show(true);
+        });
+    });
 }
 
 export async function importAndCompile(askFLags = false): Promise<any> {
@@ -76,15 +84,15 @@ export async function importAndCompile(askFLags = false): Promise<any> {
 
   const defaultFlags = config().compileFlags;
   const flags = askFLags ? await compileFlags() : defaultFlags;
-  return importFile(file, flags).catch(error => {
+  return importFile(file).catch(error => {
     console.error(error);
-  });
+  }).then(() => compile([file], flags));
 }
 
 // Compiles all files types in the namespace
 export async function namespaceCompile(askFLags = false): Promise<any> {
   const api = new AtelierAPI();
-  const fileTypes = ["*.CLS", "*.MAC", "*.INC", "*.BAS"]
+  const fileTypes = ['*.CLS', '*.MAC', '*.INC', '*.BAS'];
   if (!config('conn').active) {
     throw new Error(`No Active Connection`);
   }
@@ -94,21 +102,58 @@ export async function namespaceCompile(askFLags = false): Promise<any> {
     // User cancelled
     return;
   }
-  vscode.window.withProgress({
-    location: vscode.ProgressLocation.Notification,
-    title: `Compiling Namespace: ${api.ns}`,
-    cancellable: false
-  }, async () => {
-    const data = await api
-      .actionCompile(fileTypes, flags);
-    if (data.status && data.status.errors && data.status.errors.length) {
-      console.error(data.status.summary);
-      throw new Error(`Compiling Namespace: ${api.ns} Error`);
+  vscode.window.withProgress(
+    {
+      location: vscode.ProgressLocation.Notification,
+      title: `Compiling Namespace: ${api.ns}`,
+      cancellable: false
+    },
+    async () => {
+      const data = await api.actionCompile(fileTypes, flags);
+      if (data.status && data.status.errors && data.status.errors.length) {
+        console.error(data.status.summary);
+        throw new Error(`Compiling Namespace: ${api.ns} Error`);
+      } else {
+        vscode.window.showInformationMessage(`Compiling Namespace: ${api.ns} Success`);
+      }
+      const file = currentFile();
+      return loadChanges([file]);
     }
-    else {
-      vscode.window.showInformationMessage(`Compiling Namespace: ${api.ns} Success`);
-    }
-    const file = currentFile();
-    return loadChanges(file);
-  });
+  );
+}
+
+function importFiles(files) {
+  return Promise.all<CurrentFile>(
+    files.map(file =>
+      vscode.workspace
+        .openTextDocument(file)
+        .then(currentFile)
+        .then(file =>
+          Promise.resolve(file)
+            .then(importFile)
+            .then(data => {
+              outputChannel.appendLine('Imported file: ' + file.fileName)
+              return file;
+            })
+        )
+    )
+  )
+    .then(compile);
+}
+
+export async function importFolder(uri: vscode.Uri): Promise<any> {
+  let folder = uri.path;
+  if (fs.lstatSync(folder).isFile()) {
+    return importFiles([folder]);
+  }
+  glob(
+    '**/*.{cls,inc,mac,int}',
+    {
+      cwd: folder,
+      nocase: true
+    },
+    (error, files) => importFiles(
+      files.map(name => path.join(folder, name))
+    )
+  );
 }